主界面运动 更新Account、热量推荐算法类

+++ app/src/main/java/com/example/foodcare/model/MainFood.java
@@ -39,7 +39,7 @@ public class MainFood {
     }
 
     public double getTotalEnergy() {
-        return energyPerHectogram * foodWeight / 100;
+        return (double) energyPerHectogram * (double) foodWeight / 100d;
     }
 
     public void setFoodWeight(int foodWeight) {
